removed data processing from main thread
