+++ app/src/main/java/com/bruhascended/fitapp/ui/foodjournal/DateSeparatedItem.kt
@@ -1,5 +1,6 @@
 package com.bruhascended.fitapp.ui.foodjournal
 
+import androidx.recyclerview.widget.DiffUtil
 import com.bruhascended.db.food.entities.FoodEntry
 import java.util.*
 
@@ -7,6 +8,25 @@ class DateSeparatedItem(
     val item: FoodEntry? = null,
     val separator: Date? = null
 ) {
+
+    class Comparator : DiffUtil.ItemCallback<DateSeparatedItem>() {
+
+        override fun areItemsTheSame (
+            oldItem: DateSeparatedItem,
+            newItem: DateSeparatedItem
+        ) : Boolean {
+            return if (oldItem.isSeparator && newItem.isSeparator) {
+                oldItem.separator == newItem.separator
+            } else {
+                oldItem.item?.entry?.entryId == newItem.item?.entry?.entryId
+            }
+        }
+
+        override fun areContentsTheSame(
+            oldItem: DateSeparatedItem,
+            newItem: DateSeparatedItem
+        ) = oldItem == newItem
+    }
 
     val isSeparator: Boolean
         get() = separator != null
